Use the form without a data class
The abuse of the 'folder' field for uploaded files led to type mismatch:
the field was a string and an array at the same time.

When using the 'folder' as a fallback for 'displayName', this backfired:
the processing of the form tried to read from the display name, but read
from 'folder' instead, which already held an array, and the type check
failed.

Let us not abuse fields any more.